Refactor/move insomnia fetch (#7344)
* remove double slash

* move to ui

* use insomniaFetch where possible

* fix lint
