+++ packages/insomnia/src/models/helpers/project.ts
@@ -1,7 +1,7 @@
 import { database } from '../../common/database';
-import { insomniaFetch } from '../../main/insomniaFetch';
 import { initializeLocalBackendProjectAndMarkForSync, pushSnapshotOnInitialize } from '../../sync/vcs/initialize-backend-project';
 import { VCS } from '../../sync/vcs/vcs';
+import { insomniaFetch } from '../../ui/insomniaFetch';
 import { invariant } from '../../utils/invariant';
 import { isDefaultOrganizationProject, Project, update as updateProject } from '../project';
 import { Workspace } from '../workspace';
