+++ packages/insomnia/src/sync/git/github-oauth-provider.ts
@@ -1,6 +1,7 @@
 import { v4 } from 'uuid';
 
-import { getApiBaseURL, getAppWebsiteBaseURL, getGitHubGraphQLApiURL } from '../../common/constants';
+import { getAppWebsiteBaseURL, getGitHubGraphQLApiURL } from '../../common/constants';
+import { insomniaFetch } from '../../ui/insomniaFetch';
 
 export const GITHUB_TOKEN_STORAGE_KEY = 'github-oauth-token';
 export const GITHUB_GRAPHQL_API_URL = getGitHubGraphQLApiURL();
@@ -42,20 +43,16 @@ export async function exchangeCodeForToken({
     );
   }
 
-  const apiURL = getApiBaseURL();
-
-  return window.main.axiosRequest({
-    url: apiURL + '/v1/oauth/github',
+  return insomniaFetch<{ access_token: string }>({
+    path: '/v1/oauth/github',
     method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
     data: {
       code,
     },
-  }).then(result => {
+    sessionId: '',
+  }).then(data => {
     statesCache.delete(state);
-    setAccessToken(result.data.access_token);
+    setAccessToken(data.access_token);
   });
 }
