+++ packages/insomnia/src/sync/vcs/vcs.ts
@@ -9,8 +9,8 @@ import * as crypt from '../../account/crypt';
 import * as session from '../../account/session';
 import { generateId } from '../../common/misc';
 import { strings } from '../../common/strings';
-import { insomniaFetch } from '../../main/insomniaFetch';
 import { BaseModel } from '../../models';
+import { insomniaFetch } from '../../ui/insomniaFetch';
 import Store from '../store';
 import type { BaseDriver } from '../store/drivers/base';
 import compress from '../store/hooks/compress';
