+++ packages/insomnia/src/ui/components/toast.tsx
@@ -10,8 +10,8 @@ import {
   getProductName,
   updatesSupported,
 } from '../../common/constants';
-import { insomniaFetch } from '../../main/insomniaFetch';
 import * as models from '../../models/index';
+import { insomniaFetch } from '../../ui/insomniaFetch';
 import imgSrcCore from '../images/insomnia-logo.svg';
 import { useRootLoaderData } from '../routes/root';
 import { Link } from './base/link';
