+++ packages/insomnia/src/ui/routes/mock-route.tsx
@@ -5,12 +5,12 @@ import { LoaderFunction, useFetcher, useParams, useRouteLoaderData } from 'react
 import { CONTENT_TYPE_JSON, CONTENT_TYPE_OTHER, CONTENT_TYPE_PLAINTEXT, CONTENT_TYPE_XML, CONTENT_TYPE_YAML, contentTypesMap, getMockServiceURL, RESPONSE_CODE_REASONS } from '../../common/constants';
 import { database as db } from '../../common/database';
 import { getResponseCookiesFromHeaders } from '../../common/har';
-import { insomniaFetch } from '../../main/insomniaFetch';
 import * as models from '../../models';
 import { MockRoute } from '../../models/mock-route';
 import { MockServer } from '../../models/mock-server';
 import { Request, RequestHeader } from '../../models/request';
 import { Response } from '../../models/response';
+import { insomniaFetch } from '../../ui/insomniaFetch';
 import { invariant } from '../../utils/invariant';
 import { Dropdown, DropdownButton, DropdownItem, ItemContent } from '../components/base/dropdown';
 import { PanelContainer, TabItem, Tabs } from '../components/base/tabs';
